% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test}
\alias{test}
\title{test}
\usage{
test()
}
\description{
This test will run fine on all other architectures, including local windows,
but run forever without timing out when run on GH-Actions Windows builder.
}
\examples{
test()
}
